Fixes AppScore, Standardize IPA Binary rules
