+++ StaticAnalyzer/views/android/code_analysis.py
@@ -16,9 +16,11 @@ from StaticAnalyzer.views.android.rules import (
     android_rules,
 )
 from StaticAnalyzer.views.shared_func import (
+    url_n_email_extract,
+)
+from StaticAnalyzer.views.rule_matchers import (
     api_rule_matcher,
     code_rule_matcher,
-    url_n_email_extract,
 )
 
 logger = logging.getLogger(__name__)
