+++ templates/pdf/ios_report.html
@@ -321,39 +321,39 @@
                       </tr>
                   </thead>
                      <tbody>
-                   {% for bin in binary_analysis %}
+                   {% for issue, details in binary_analysis.items %}
                     <tr>
                       <td>
-                        {{ bin.issue }}
+                        {{ issue }}
                       </td>
                       <td>
-                        {% if bin.level == 'info' %}
-                          <span class='info'>{{bin.level}}</span>
-                        {% elif bin.level == 'good' %}
-                          <span class='success'>{{bin.level}}</span>
-                        {% elif bin.level == 'high' %}
-                          <span class='danger'>{{bin.level}}</span>
-                        {% elif bin.level == 'warning' %}
-                          <span class='warning'>{{bin.level}}</span>
+                        {% if details|key:"level" == 'info' %}
+                          <span class='info'>info</span>
+                        {% elif details|key:"level" == 'good' %}
+                          <span class='success'>secure</span>
+                        {% elif details|key:"level" == 'high' %}
+                          <span class='danger'>high</span>
+                        {% elif details|key:"level" == 'warning' %}
+                          <span class='warning'>warning</span>
                         {% endif %}
                       </td>
                       <td>
                         <h5>CVSS V2:</h5>
-                        {% if bin.cvss > 6 %}
-                        <span class="danger">{{ bin.cvss }} (high)</span>
-                        {% elif bin.cvss == 0 %}
-                        <span class="info">{{ bin.cvss }} (info)</span>
-                        {% elif bin.cvss >= 4 %}
-                        <span class="warning">{{ bin.cvss }} (medium)</span>
-                        {% elif bin.cvss < 4 %}
-                        <span class="success">{{ bin.cvss }} (low)</span>
+                        {% if details|key:"cvss" > 6 %}
+                        <span class="danger">{{ details|key:"cvss" }} (high)</span>
+                        {% elif details|key:"cvss" == 0 %}
+                        <span class="info">{{ details|key:"cvss" }} (info)</span>
+                        {% elif details|key:"cvss" >= 4 %}
+                        <span class="warning">{{ details|key:"cvss" }} (medium)</span>
+                        {% elif details|key:"cvss" < 4 %}
+                        <span class="success">{{ details|key:"cvss" }} (low)</span>
                         {% endif %}
-                        {% if bin.cwe %}</br> <h5>CWE: </h5>{{bin.cwe}}{% endif %}
-                        {% if bin.owasp %}</br> <h5>OWASP Top 10: </h5>{{bin.owasp}}{% endif %}
-                        {% if bin|key:"owasp-mstg" %}</br> <h5>OWASP MASVS: </h5>{{bin|key:"owasp-mstg"}}{% endif %}
+                        {% if details|key:"cwe" %}</br> <h5>CWE: </h5>{{ details|key:"cwe" }}{% endif %}
+                        {% if details|key:"owasp" %}</br> <h5>OWASP Top 10: </h5>{{ details|key:"owasp" }}{% endif %}
+                        {% if details|key:"owasp-mstg" %}</br> <h5>OWASP MASVS: </h5>{{ details|key:"owasp-mstg" }}{% endif %}
                        </td>
                       <td>
-                        {{ bin.description}}
+                        {{ details|key:"detailed_desc" }}
                       </td>
                     </tr>
                     {% endfor %}
