+++ templates/static_analysis/ios_binary_analysis.html
@@ -572,40 +572,40 @@
                       </tr>
                   </thead>
                      <tbody>
-                   {% for bin in binary_analysis %}
+                    {% for issue, details in binary_analysis.items %}
                     <tr>
                       <td>
-                        {{ bin.issue }}
+                        {{ issue }}
                       </td>
                       <td>
-                        {% if bin.level == 'info' %}
-                          <span class='badge bg-info'>{{bin.level}}</span>
-                        {% elif bin.level == 'good' %}
-                          <span class='badge bg-success'>{{bin.level}}</span>
-                        {% elif bin.level == 'high' %}
-                          <span class='badge bg-danger'>{{bin.level}}</span>
-                        {% elif bin.level == 'warning' %}
-                          <span class='badge bg-warning'>{{bin.level}}</span>
+                        {% if details|key:"level" == 'info' %}
+                          <span class='badge bg-info'>info</span>
+                        {% elif details|key:"level" == 'good' %}
+                          <span class='badge bg-success'>secure</span>
+                        {% elif details|key:"level" == 'high' %}
+                          <span class='badge bg-danger'>high</span>
+                        {% elif details|key:"level" == 'warning' %}
+                          <span class='badge bg-warning'>warning</span>
                         {% endif %}
                       </td>
                       <td>
                         <strong>CVSS V2:</strong>
-                        {{ bin.cvss }}
-                        {% if bin.cvss > 6 %}
+                        {{ details|key:"cvss" }}
+                        {% if details|key:"cvss" > 6 %}
                           (high)
-                        {% elif bin.cvss == 0 %}
+                        {% elif details|key:"cvss" == 0 %}
                           (info)
-                        {% elif bin.cvss >= 4 %}
+                        {% elif details|key:"cvss" >= 4 %}
                           (medium)
-                        {% elif bin.cvss < 4 %}
+                        {% elif details|key:"cvss" < 4 %}
                           (low)
                         {% endif %}
-                        {% if bin.cwe %}</br> <strong>CWE: </strong>{{ bin.cwe }}{% endif %}
-                        {% if bin.owasp %}</br> <strong>OWASP Top 10: </strong>{{ bin.owasp }}{% endif %}
-                        {% if bin|key:"owasp-mstg" %}</br> <strong>OWASP MASVS: </strong>{{ bin|key:"owasp-mstg" }}{% endif %}
+                        {% if details|key:"cwe" %}</br> <strong>CWE: </strong>{{ details|key:"cwe" }}{% endif %}
+                        {% if details|key:"owasp" %}</br> <strong>OWASP Top 10: </strong>{{  details|key:"owasp" }}{% endif %}
+                        {% if details|key:"owasp-mstg" %}</br> <strong>OWASP MASVS: </strong>{{  details|key:"owasp-mstg" }}{% endif %}
                       </td>
                       <td>
-                        {{ bin.description}}
+                         {{ details|key:"detailed_desc" }}
                       </td>
                     </tr>
                     {% endfor %}
